+++ cloudshell/cli/service/command_mode_helper.py
@@ -2,7 +2,6 @@ from collections import OrderedDict
 
 from cloudshell.cli.service.command_mode import CommandMode, CommandModeException
 from cloudshell.cli.service.node import NodeOperations
-from cloudshell.cli.session.session import Session
 from functools import reduce
