+++ tests/cli/session/test_telnet_session.py
@@ -12,7 +12,7 @@ class TestTelnetSession(TestCase):
         self._port = 22
         self._on_session_start = Mock()
 
-    @patch('cloudshell.cli.session.telnet_session.ExpectSession')
+    @patch('cloudshell.cli.session.session.AbstractSession')
     @patch('cloudshell.cli.session.telnet_session.ConnectionParams')
     def test_init_attributes(self, connection_params, expect_session):
         self._instance = TelnetSession(self._hostname, self._username, self._password, port=self._port,
@@ -20,7 +20,7 @@ class TestTelnetSession(TestCase):
         mandatory_attributes = ['username', '_handler', 'password']
         self.assertEqual(len(set(mandatory_attributes).difference(set(self._instance.__dict__.keys()))), 0)
 
-    @patch('cloudshell.cli.session.ssh_session.ExpectSession')
+    @patch('cloudshell.cli.session.ssh_session.AbstractSession')
     def test_eq(self, expect_session):
         self._instance = TelnetSession(self._hostname, self._username, self._password, port=self._port,
                                        on_session_start=self._on_session_start)
